Restructure the python testing directory to be more extensible and use simple annoying_hack.py to make sure we can import utils (now cistem_test_utils) from horizontal level directories (now named after parent program, eg match_template.
