+++ scripts/testing/programs/cistem_test_utils/compare_coordinates.py
