+++ scripts/testing/programs/match_template/test_fringe_image.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python3
 
+import annoying_hack
 
 from os.path import join as join
-import util.args as tmArgs
-import util.make_tmp_runfile as mktmp
-import util.run_job as runner
-import util.re_run_results_on_mip as re_runner
-import sys
+import cistem_test_utils.args as tmArgs
+import cistem_test_utils.make_tmp_runfile as mktmp
+import cistem_test_utils.run_job as runner
+import cistem_test_utils.re_run_results_on_mip as re_runner
+
 
 # By default the "_gpu" suffix will be added unless the --old-cistem flag is used
 # or the --cpu flag is used
@@ -24,7 +25,7 @@ def main():
 
     # 0 is the fring image
     # 1 is the all fun stuff image
-    config = tmArgs.get_config(args, 'Lamella_from_je', 1, 1)
+    config = tmArgs.get_config(args, 'Lamella_from_je', 0, 0)
 
     tmp_filename_match_template, tmp_filename_make_template_results = mktmp.make_tmp_runfile(config)
     elapsed_time = runner.run_job(tmp_filename_match_template)
